Task 3

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -12,7 +12,8 @@ int set_bit(unsigned long int *n, unsigned int index)
 	{
 		return (-1);
 	}
-	unsigned int mask = 1 << index;
+
+	unsigned long int mask = 1 << index;
 
 	*n |= mask;
 	return (1);
